+++ src/app/how-to-contribute/how-to-contribute.component.tsx
@@ -1,4 +1,10 @@
 import React from 'react';
 import { Header } from './header/header.component';
+import { Description } from './description/description.component';
 
-export const HowToContribute = () => <Header />;
+export const HowToContribute = () => (
+  <>
+    <Header />
+    <Description />
+  </>
+);