+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-#  (2020-08-20)
+#  (2020-08-27)
 
 
 ### Features
 
 - About description component ([4a45235](https://github.com/Exlabs/value-driven-templates/commit/4a45235f3c3411154403c5f1291df0d4a1fba25a))
+- Contribution Description ([3b1ee69](https://github.com/Exlabs/value-driven-templates/commit/3b1ee694a6aff944d5fbd55b838ea32f1f9d1fa4))
+- Contribution step ([61a3280](https://github.com/Exlabs/value-driven-templates/commit/61a3280427d096a01ebb83d31743d04b6b8f7f9e))
 - Footer component ([b10f707](https://github.com/Exlabs/value-driven-templates/commit/b10f707e635e773ee90fbb0a938abac715fc23e5))
 - Header about component ([fba1867](https://github.com/Exlabs/value-driven-templates/commit/fba18675d68e81e4e6a3e56131c13976435407f6))
 - Header component ([1d4c357](https://github.com/Exlabs/value-driven-templates/commit/1d4c3570202c8eef6bf1b93c251d6d8395abf106))
+- How to contribute header ([14a2e01](https://github.com/Exlabs/value-driven-templates/commit/14a2e014464059afacf63fe1c00f02ac1f2b317b))
 - Mobile menu component ([a3eb900](https://github.com/Exlabs/value-driven-templates/commit/a3eb900d67a5ffbb29f759b0e6d20fcaad38cadd))
 - Navigation component ([498e4b9](https://github.com/Exlabs/value-driven-templates/commit/498e4b977be4cd96f56898b99094c4b90f32495c))
 - Tag component ([5c874db](https://github.com/Exlabs/value-driven-templates/commit/5c874db1a12da5f625ed77f82f4b9cb52cf112bb))
@@ -18,7 +21,7 @@
 ### Refactoring & Improvements
 
 - Added logo to navigation ([3791328](https://github.com/Exlabs/value-driven-templates/commit/37913282b9f0c94ec09ee7e5e1ad0bf0bf997865))
-- Align improvements ([044d4c1](https://github.com/Exlabs/value-driven-templates/commit/044d4c1552885c2aa5c878164d8b8c16fc915d8a))
+- Align improvements ([ea6e1c4](https://github.com/Exlabs/value-driven-templates/commit/ea6e1c4dd627a542b176387a8d0c4b4ec14e4485))
 - Deployment improvement. Added static files prefix. ([653750b](https://github.com/Exlabs/value-driven-templates/commit/653750b2629faa9adbffb5e0557a18f3ad409258))
 - Empty thumbnail list improvement ([2148cb3](https://github.com/Exlabs/value-driven-templates/commit/2148cb37f7b1c79065779ffa990a7402b0d93227))
 - Fixed sticky nav jumping on mobile ([f0d3f87](https://github.com/Exlabs/value-driven-templates/commit/f0d3f87e6cd8704b03de76a276cd0fd237c95f78))